FEATURE: Associate OpenIDs with an existing account
This commit finishes the end-to-end basic user experience - you can now
add an OpenID to an existing account, and login using that ID.
Add/remove should work properly, as well as handling edge cases like
cancelling associating an OpenID, trying to add an OpenID twice, etc.

QA: Have at it now. Note there are still a number of feature stories
that aren't written yet (particularly, signing up using an OpenID, but
there are others) so be sure to look at the openid tag in Pivotal before
considering a lack of feature to be a bug.

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>