NOOP: Clean up exception handling and printing
I had some ugly code that was removing the "Validation failed:" portion
of a core server error - I figured it's easier to just change it
server-side to not display that part.

QA: You might notice minor differences in error messages you can see,
but mostly this is too low-level to notice at the UI level. The
particular area I coded this for was handling the error that gets
displayed when you try to add an OpenID but it fails to do so, such as
when you add the same OpenID twice.

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>