BUG: Linking OpenIDs to an existing account should always start on "Select a provider" page
If you've logged in with OpenID before, RPX maintains a cookie so
popping up the login box again says something like "Welcome back" and
pre-selects the last provider you used. Useful for login, confusing for
when you're adding a new OpenID. This checkin overrides it on the login
only so that you start "fresh".

QA: Login with OpenID, then try to add a new OpenID. The popup should
display the list of providers, rather than defaulting you to the one you
signed in with.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>