NOOP: When trying to authenticate with RPX, create stub OpenIdIdentifiers if we don't recognize the OpenID
RPX profile tokens are only valid for 10 minutes, so in order to sign up
a new user that signed in via OpenID, I need to store the identifier
somewhere more permanently. I decided to change the core server to
create an OpenIdIdentifier, and pass it back as an undocumented field on
the fake User payload to Rails.

The net effect of this is that you can now create a identifier that's
not tied to a user.

QA: Nothing to test

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>