FEATURE: Login via OpenID with unrecognized account leads to signup
From this checkin forward, signing in with an OpenID that's not
currently linked to a Socrata account will present a customized version
of the signup screen with as much as we could get from the OpenID
provider pre-filled-in. It also has a different explanatory blurb about
us needing to create a socrata account, and how to link an OpenID to an
existing account by logging in and visiting the My Account page.

This still doesn't complete the usability improvements - I want to also
add an option to signin with Socrata credentials right from this page to
sign them in and link the credentials w/o having them visit the My
Account page. That's not done.

QA: Testable. Be sure to check both forms of the signup form (the one
you normally visit, and the one that comes up by signing in with an
unrecognized OpenID.) You'll also want to try a few different OpenID
providers (I recommend MyOpenID) to see what profile information is
carried over into our signup form.

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>