BUG: Some users can't really do anything in the app (Request forgery protection/invalid cookie)
This is a further application of b98064021d791ddad351c066b00ba3d296833620
which apparently is the bug that refuses to die. I was incorrect in my
original assesment that only Rails 2.0.x got the encoding wrong and
needed a patch; it turns out that Rails 2.3.x still has the same
problem.

Applying a nasty hack to fake URL-encode the string before URL decoding
it fixes the problem in the web/ project.

QA: http://www.pivotaltracker.com/story/show/1210031
    http://www.pivotaltracker.com/story/show/1210188

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>