BUG: Fix bad interaction between JSON gem and Rails 2.3.3 JSON support
We were seeing a variety of issues using JSON.generate - see
http://groups.google.com/group/rubyonrails-talk/browse_thread/thread/3739dd637e038fa9
for details of what's wrong and why I made this fix.

QA: This was the hot-patch I applied directly to test.socrata.com on
Monday night. It solves most (but not all) of the Something's Wrong
errors you saw over the weekend.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>