BUG: Updating a user profile with no tags doesn't close edit section or update until refresh
QA: http://www.pivotaltracker.com/story/show/1218471, but it mistakenly
marked this bug as IE7-specific. That's not the case; it doesn't work in
any browser.

It seems like http://www.pivotaltracker.com/story/show/1160736 may be
related, but it's certainly not the same bug. I'd re-test it to see if
it's still happening.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>