NOOP: Refactor signup logic into a Presenter class
In a change in the near future, I'm going to need the ability to create
a user account from more than one controller; the current code in
AccountsController#create is really large and ugly, doing a bunch of
things at once:
* Creating the user account
* Logging in as the newly-created user
* Optionally setting their profile picture, if provided
* Optionally sending invite emails to others, if provided

Rather than pollute the controller with all that logic (and gobs of
error handling) I made a SignupPresenter that encompasses all of that
logic. As a bonus, it cleans up our form handling code for the signup,
since all the fields can now be part of the presenter, so we no longer
need to intermix *_field and *_field_tag helpers in the same form.

QA: There's a pretty high risk of regression on the signup path. I'd
specifically test that validation errors continue to work as expected,
that it handles invites and profile photos correctly, and that all the
optional fields on the profile continue to work.

Reviewed-by: Sam Gibson <sam.gibson@socrata.com>