BUG: OpenID signup wasn't properly setting OpenID profile info
I still don't understand why flash[:foo] = ...; redirect_to(...) was
working before but stopped working a few days ago - this is my 2nd or
3rd bug with the same thing.

This code is going to be rewritten soon into a new controller where
we'll save a redirection and clean up the account creation path. In the
meantime, I hacked around this by using the session instead of the
flash.

QA: http://www.pivotaltracker.com/story/show/1250328

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>