BUG: Fix flash hash support, and a bunch of other random things that might be broken.
The custom benchmarking code I wrote had a very nasty, difficult to
track down bug in it - we were skipping some functionality from the core
Rails, which could change from release to release.

https://rails.lighthouseapp.com/projects/8994/tickets/285-alias_method_chain-limits-extensibility
explains the bug (and notes that it's a wontfix for the 2.x series in
Rails, boo!) but the gist of it is that we were skipping some logic on
perform_action, which among other things saved the flash hash to the
session on each request. Read the comment in the code if you want the
gory details.

QA: All the spots that had missing validation messages, permission
denied messages, login/logout confirmations, etc. should be fixed now.
http://www.pivotaltracker.com/story/show/1251364
http://www.pivotaltracker.com/story/show/1250220
http://www.pivotaltracker.com/story/show/1251545

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>