FEATURE: Allow the user to switch between linking an OpenID to an existing account and creating a new account
This checkin introduces the UI for logging in with an OpenID first, then
choosing whether to link it with an existing Socrata account or create a
brand-new account. Both options will create the link between the OpenID
and the account.

At this point, the visuals are complete, but the actual functionality is
still somewhat broken.

QA: Do not test

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>