FEATURE: Using an unrecognized OpenID gives a combined signup/login experience
Finally, it's here! If you sign in with an OpenID that we don't
recognize, we have a custom screen that tells you we don't recognize it,
and offers the ability to either create a new account or login with an
existing account, either of which ties the OpenID to it.

QA: http://www.pivotaltracker.com/story/show/1096738

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>