BUG: Fixed a number of spots where the validation errors continued to show past their useful lifetime
In rails, a flash[:foo] assignment lasts for the duration of the current
request and the next request. It's usually a bug if you have something
like:
  flash[:error] = some_error
  return (render :template => 'shared/error')

This means that the flash survives one extra page beyond what's
necessary, and if it's something like the login page, it might not make
any sense to put it there. There are two different fixes:

1. Use flash.now instead of flash, so it only lasts the lifetime of the
current request. (Most fixes in this commit follow that model)
2. Use redirect_to instead of render, so you're actually redirecting
them to the relevant page. This makes more sense when you don't want to
actually do things like preserve an invalid form submission.

QA: http://www.pivotaltracker.com/story/show/1268533
    http://www.pivotaltracker.com/story/show/984103

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>