BUG: Cleanup session usage for OpenID identifiers during signup/login
Up until now, I've been using the session as a cheap, secure way to
ensure that OpenID identifiers aren't tampered with. The problem stems
from the fact that OpenID identifiers are monotonically increasing
integers, and we can't trust the user to not tamper with hidden fields
on the form to grab other user's OpenID identifiers during the signup
process by repeatedly trying to register the next OpenID. (Granted,
there's not much they can do with it other than become an annoyance, but
it's an attack vector nonetheless.)

To solve this, I chose to actually make the IDs of the OpenID
identifiers opaque and digitally signed - now, it's okay to put the
identifiers in the form, since it's not feasible to create a different
identifier without knowing the secret.

This had weird side-effects like Roopa noticing that the login code
could get "stuck" if you somehow had a session with an invalid OpenID
identifier; that bug is fixed as a side-effect of this change.

QA: http://www.pivotaltracker.com/story/show/1269139
You'll want to give all the signup/login code a good workout, however -
this was a pretty far-reaching change.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>