BUG: Uniquify cache name for rpx/return_token javascript
I copy-pasted some code when refactoring from the normal signup page,
and forgot to change it to use a different javascript cache file. Weird
things break when running in production...

QA: This is the bug I patched last night on the server to make the
OpenID signup/login stuff work again.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>