BUG: Ajax upload sometimes not following the mouse properly (popup login profile pictures weren't working)
This was a really hard-to-track-down bug. Essentially, the Ajax upload
control sometimes worked and sometimes didn't; the only way to really
figure out what was happening was to edit the plugin to bump up the
opacity of the "hidden" file upload control that follows the mouse
around to see what was happening.

The easiest reproduction steps work something like this:
1. Bring up the dialog with the upload control
2. Select some other top-level window in your browser to give it
mouse/keyboard focus.
3. Now you can move your mouse around over the browser window without
any mouse movement events firing. Move the mouse directly over the
upload control, and click to select it (On OSX, you'll need to click a
2nd time since the first click only gives the window focus)
4. This will load the wrong (non-Ajax) file upload control's browse
dialog. Weird stuff happens from there - the file doesn't look like it
gets uploaded, the wrong validation routines run, etc.

QA: The situation above should give you a pretty good idea of how to
test this. Another way to ensure you lose mouse focus is to use a
javascript debugger like Firebug, with a breakpoint in the onChange()
function in the AjaxUpload($('#signup #signup_profile_image'))
constructor.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>