FEATURE: Profile page has help text hinting at what should go in missing fields
This change implements part of a mockup from Mark that's part of a
greater change to make our signup process ask for less data, but on
completion send you to the profile page to fill out any extra stuff they
feel like adding. In particular, this change only makes it more obvious
what our users should be adding (company, title, location, tags, about
me, interests, and external links) and makes it easier to discover how
to edit each section when they're not filled out.

QA: Edit data on the profile page. The experience should be essentially
the same, but if you leave some data out, expect some italicized,
grey-colored text hinting that you should add some.

One known issue: IE7 no longer displays the prefix text "Tags:" before
the list of tags. IE7 doesn't support CSS before elements, and in this
case it's graceful degradation, so it's not a bug we're going to fix.
Don't file a bug on that.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>