FEATURE: New buttons on the profile and friends pages
This checkin adds some buttons to make navigation easier:
* The Profile page gets "Design a Dataset", "Find Friends", and "Invite
  Others" buttons. They're only valid when you're viewing your own
  profile, not someone else's.
* The Friends tab gets two buttons "Find Friends" and "Invite Others"

The former page is based on a spec from Mark; the latter is because I
felt like it and was really annoyed that the friends page offered no way
to add new friends once you've added a single one.

There's also one variation from Mark's spec: the profile page is
supposed to list the number of invites sent; while we track that info I
think it's kind of useless to display to the user and not worth the cost
of an additional query and bigger JSON payload for User objects. It
looked funny putting two buttons next to each other, so I put some
generic blurb text about inviting others to make it look nice.

QA: Make sure the new buttons described above work.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>