FEATURE: Redesign the signup and login screens to ask for less information
As of this checkin, all we ask for during signup is a username,
password, and email address. (Password will go away for OpenID signups
soon - but not in this checkin.) There are a lot of visual changes
you'll want to reference Mark's mockups to get a good idea of the extent
of them.

QA: There are a number of cases you'll want to test:
* Normal login
* Normal signup
* OpenID signup from the login page and from the signup page
* OpenID account linking from the login page and the signup page
* Login from the popup
* Sign up from the popup (note you can't use OpenID here because OpenID
  would redirect the browser, killing the point of the AJAX popup)

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>