BUG: RPX popup wasn't loading on the signup page
I forgot to load the javascript necessary for RPX to function... it was
still only on the login page, rather than the signup page. Since this is
the signup page, I also set the flag that tells RPX to ignore any
previously selected login provider on this page, so you'll always be
presented with the list of 6 options here.

QA: http://www.pivotaltracker.com/story/show/1329244

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>