FEATURE: Attempting to remove last OpenID w/o password ensures password section is open
As a hint on what you need to do to set a password, we helpufully open
the "Create password" section on the My Account page if you try to
remove your last OpenID when you have no password set.

This change also refactors a large amount of the expand/collapse code
into a jQuery UI-style widget; all the code responsible for
expand/collapsing the sections on this page are effectively re-done.

QA: Self-explanatory

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>
Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>