NOOP: Fully convert profile and my account pages to use new show/edit plugin
There shouldn't be any change in functionality here; I'm just changing
the code that toggles between the show and edit versions of the editable
sections on the profile and my account pages to use the plugin.

QA: Make sure the expand/collapse of the show/edit sections on both
pages still work properly.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>