BUG: Fix HTML escaping of description in header [SECURITY]
Although I wasn't able to get exploitable code of it, I was able to
screw up the page rendering by setting the description of a dataset to:
"/>I shouldn't be here!

It'd result in the text showing up above all the content on the dataset
host page or inside the widget. In theory you might be able to get it to
run arbitrary Javascript, but I wasn't able to get that to happen.
Regardless, this fixes the nasty visual bug.

QA: Try setting a dataset to have the description:
"/>I shouldn't be here!
and reload. You might also try other partial-HTML tags, etc. to see if
you can get it to mess up the rendering of the page.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>