FEATURE: Delay-load tab content in widget
Using the net tab in firebug, I noticed that even though we agressively
cache the widget page in production, most of that gain is lost by the
fact that we were making 8 (!) requests to fill in the dynamic bits we
needed for each tab.

We still make those requests, but now we only do so if you switch to one
of those tabs. You might notice a small delay for the content on a
widget tab to pop into display, but it should visually look the same
once it's loaded.

QA: Self-explanatory

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>