SPARTA: Performance speedup loading widgets (views.filters.length)
Profiling our requests for a widget, I noticed a particular request to
the core server that seemed odd:

Core server request: /views.json?method=getByTableId&tableId=xxx (yy.zms)

This request was happening because we were rendering the partial
blists/info_filtered_header, which calls view.filters.length.
view.filters makes a request of the core server to get all the views
associated with a dataset; all we wanted is to display the string
"n Filtered Views". Worse, this was being called when rendering the
widget initially AND as a call when loading the "More Views" tab as an
XMLHttpRequest.

I took the opportunity to refactor this a bit:
* Rendering various tabs as an XMLHttpRequest for the widget will first
  look up whether a partial exists in app/views/widgets, and fall back
  to rendering the view in app/views/blists. This means if we have
  significantly different logic between the two, we can override it
  entirely rather than switching based on an is_widget local variable.
* I changed the widget version of the info_filtered_header partial to
  simply render static HTML with an empty span for the number of
  filtered views and some CSS magic to make the string "Filtered Views"
  plural or not. There's some javascript that runs when the widget
  retrieves that section to insert the view count based on the table of
  views (which came as a separate request).

The net effect is that you might occasionally see the text "Filtered
Views" without a count, but once we have the actual table of filtered
views displayed, the count should be accurate.

QA: Check that the "More views" tab in the widget and dataset page
continue to function as expected.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>