SPARTA: Cache <noscript> HTML representation of widget rows in memcache
The slowest part of our request for a widget at this point is retrieving
an bare HTML representation of the rows, which we show in the widget
inside a <noscript> tag for SEO/accessibility. The rows are retrieved a
2nd time in a richer form that real viewers see.

This checkin caches that HTML representation for an hour. Changes to
data on the widget might appear stale to search engines, but then again
they don't tend to come around every few minutes so that's not a big
deal. Active widgets (accessed more often than once/hour) are now
making only 2 requests to the core server at steady state - the view
JSON and the user JSON.

QA: Widgets should still load. If you inspect the HTML of the widget
IFRAME, you'll still see the row data inside the <noscript> section.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>