SPARTA: Remove static page caching for widgets
The static page caching we've been doing wasn't going to last much
longer, with the introduction of premium SDPs we needed a way to
invalidate the cached version. All the earlier optimizations have the
request times down to about 100ms - at this point, it seems reasonable
to just let Rails process it now.

* Remove page cache
* Remove workaround for getting user_id as a 2nd request, now that we
  can do custom logic on each request.
* Added a deprecation notice via Hoptoad for if someone calls the
  workaround for getting the user_id - I don't think anyone is using
  it anymore, but better to leave it in and find out.

QA: Widgets should continue to function, and remain performant.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>