FEATURE: Report Javascript errors to the Hoptoad
This checkin introduces an ErrorsService on the core server, which can
be POSTed to with some parameters that constructs a basic Hoptoad notice
to the core server.

In Javascript, I've written a few wrapper functions that catch most
common error cases in Javascript in a try/catch block, allowing us to
intercept the exception and (depending on the Rails environment) display
an alert, report the error to hoptoad, or both:
* In development, no errors are logged to the server, but you will get a
  popup alert for any errors we catch. Due to irregularities in the way
  can catch errors, Webkit-based browsers will often catch fewer errors,
  and some other browsers will double-report errors.
* In staging (test.socrata.com) errors cause a popup AND report it to
  the server.
* In production, no error is displayed to the client but we report it to
  the server.

To test this, I created a new controller in Rails that's only exposed in
non-production environments - visit /test_page to see a list of test
pages, or in this specific case, you can visit /test_page/js_kaboom.

QA: If you visit /test_page/js_kaboom, you should trigger two different
errors on load, and pressing the button will trigger a 3rd. Webkit-based
browsers (Chrome, Safari) will likely report fewer; that's not a bug in
our code, it's a bug in those browsers we can't fix. You should be able
to find the errors you report on the "JS Client" project at
https://socrata.hoptoadapp.com/projects/7928/errors

Do not report bugs about the styling of these test pages. They're not
exposed in production, so they're not getting a level of polish to
match. If we have some free time we'll go through and try to make them
nicer, but it's super-low-priority.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>