FEATURE: Asset package resources
This checkin actually makes use of the asset packager plugin. In order
to deploy, we'll have to add a step to the Rails deployment that runs
rake asset:packager:build_all - ideally, on a single machine and then
rsyncing the contents to the other hosts.

QA: This change reorganized almost every javascript and CSS file on the
site, and re-introduced a minification routine. This has a very high
probability of regressions where a feature on some page won't work,
either because I messed up the javascript includes or a latent bug in our
Javascript caused it to not minify properly. This needs a broad sweep of
the entire site.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>