BUG: Fix broken Javascript that made YUICompressor choke
Yes folks, someone decided that int was a good name for a variable.
It's technically not used in Javascript, and debatable as to whether
it's a reserved word. Regardless, YUICompressor chokes on it, so let's
change it to a reasonable variable name.

QA: This should fix the discover and community pages.

Reviewed-by: Clint Tseng <clint.tseng@socrata.com>