BUG: Error handling for Append/Replace dialogs
The error handling for this dialog was nearly nonexistent, and basically
crashed the javascript app if you uploaded something invalid. I'm not
claiming we'll get USEFUL error messages now, but we will display
whatever it is the core server decides to tell us.

QA: http://www.pivotaltracker.com/story/show/1441539
    http://www.pivotaltracker.com/story/show/1441584

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>