FEATURE: Setup favicon.ico access for brain-dead clients
Until now, we were specifying a meta tag in our HTML to direct browsers
to use an appropriately-themed favicon. That worked for smart browsers,
but a lot of other browsers and bots still wanted to just pull the icon
hardcoded from /favicon.ico.

Rather than try to be smart in the HTML, let's just be smart when
returning the response at favicon.ico - read the file from the
appropriate theme, otherwise return a 404 Not Found.

QA: You should see an appropriate favicon for the Socrata, gov, and ohioag
themes. Modern browsers display it near the URL, and you can always
retrieve it by going to /favicon.ico.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>