BUG: Non-existent pages were returning 500 instead of 404 errors
If you visit a random, non-existent URL, you'd notice we've lost our
friendly "Not found" page and it's just a plain-old white page;
inspecting with Firebug shows it's returnsing HTTP 405 (Method not
allowed) and we're getting Hoptoad errors for it, when we shouldn't.

This change fixes the patch I did early for Microsoft Office clients
(0f5b9784e3db67a30eb60516b9c48a875f7d671f) so that Rails doesn't try to
catch-all routes and say that the only method type allowed is OPTIONS.
Instead of a route, we now handle it with a before_filter.

QA: This is inadvertently fixing
http://www.pivotaltracker.com/story/show/1459567 - my instructions on
that commit were slightly off on the URL - it's of the form
/blists/xxxx-xxxx/sort_bys, not the
/dataset/Friendly-Name/xxxx-xxxx/sort_bys URL. For the latter, you
should now see the friendly "Not Found" page instead of the white page
reported in that bug.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>