BUG: Memcache was dumb and falls over horribly when it can't connect
This investigation came about as part of diagnosing why www.socrata.com
keels over and dies when util01 reboots. To test this, I pulled one app
server out of the rotation, added an IPTables rule to block it's ability
to access util01, and then then tried accessing it directly. Looking at
the logs, it appears that we were spending an inordinate amount of time
waiting for memcache to fail.

Looking at the memcache-client code that's vendored with rails, the
problem became apparent: the old, unreliable timeout code for gets,
read, and write were replaced with a fancy new BufferedIO object that
performs nonblocking reads at the raw socket level, which works great if
you actually made it to read. However, we were failing in connect. You
can test this yourself by running the following in irb:

require 'socket'
TCPSocket.new('10.0.0.177', 11211) # Use an IP that has no machine on it
  .... seconds pass .... the grass grows an inch ....
Errno::ETIMEDOUT: Operation timed out - connect(2)
  from (irb):15:in `connect'
  from (irb):15

Clearly, we're stuck connecting the socket in the first place, which has
no real timeout connection beyond the operating system's default timeout
on TCP sockets.

The fix is to re-add the timeout support, and patch a few spots around
the code where the timeout was causing the server to try closing the
socket, which cleared the retry timer unnecessarily and put us in a loop
of each memcache operation timing out rather than waiting the dead
server RETRY_TIME.

QA: Nothing to test.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>