BUG: Timeout the blog RSS feed if it's taking too long
Last night, Mzima had a 3-hour partial outage where some of us could
connect to http://www.socrata.com, and some of us couldn't - it was
simply a function of what network you're on. (Our office couldn't
connect to the datacenter, for example.)

During this time, those of us who could connect to www.socrata.com
couldn't log in, because /home was timing out. The reason for this was
that the homepage loads the corporate RSS feed; the datacenter could
connect to blog.socrata.com, but couldn't follow the redirect to
feeds.socrata.com (aka Feedburner), which was on the other side of the
netsplit.

This change makes it so that the webapp gets at most 1 second to
retrieve the RSS feed; if it fails to do so in that timeframe - too bad,
no RSS feed for you.

QA: You can't really test this failure scenario because odds are the RSS
feed will load just fine. You should verify that in the normal case, the
RSS feed continues to show up.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>