BUG: Fix authentication issues with commons-codec 1.4
The upgrade to commons-codec 1.4 changed some details on how Base64
encoding works. In particular, the new version it adds a newline to the
normal Base64 encoding, and we switched to a URL-safe variant that
exchanges + and / for - and _, and stops appending 0-2 '=' signs to pad
the total payload to a multiple of 4 characters.

QA: You'll want to regression test login, remember me functionality,
etc. There shouldn't be any noticeable change in functionality.

Reviewed-by: Kostub Deshmukh <kostub.deshmukh@socrata.com>