BUG: XSS vulnerability in /datasets/detail?items=
The items parameter (expected to be an integer) wasn't being properly
sanitized, resulting in the ability to pass arbitrary Javascript to the
client.

While I was at it, I removed a silly partial that literally switched on
the same conditions in the controller to render three different
sub-partials, instead moving the logic directly into the controller.

QA: No functional change, but be sure the detail/summary tabs below the
dataset list at /datasets functions properly, paying specific attention
to the three different states it can be in (nothing selected, one item
selected, multiple items selected)

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>