BUG: Javascript injection via malicious HTTP referer an dataset page
IOActive noted that a malicious browser could inject Javascript into
/dataset/Friendly-Name/abcd-1234 pages via a malicious Referer header.

This was injected into the page long ago to help with our flash popups
back when we still had parts of the blist app in place; I searched
through the code for any remaining references to blistGrid.referer and
couldn't find any, so the fix was just to remove the vulnerable code.

QA: Nothing specific to test.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>