BUG: IOActive-reported blist_session_id cookie should be HttpOnly
IOActive's "juciest" vulnerability they found relied in part on the fact
that they could inject javascript into our pages that loaded a URL with
document.cookies as part of the URL; this gave them access to a session
object, enabling them to steal the credentials of any other user.

While the fix for the immediate problem is to fix the
cross-site-scripting vulnerabilities, as a defense in depth measure this
change adds the HttpOnly flag to the session cookie, making it
inaccessible in Javascript on modern (IE6SP1+, FF 2.0+, etc.) browsers.

There was one spot where we were relying on the cookie in Javascript -
the top-right header links for sign in/sign out changed dynamically
based on the cookie contents. We're now setting a different cookie
(logged_in) instead that can be safely used without HttpOnly.

Note this fix isn't complete; the core server also needs an update to
set the HttpOnly flag, otherwise you'll be vulnerable if the last
request was to a core server endpoint (like retrieving rows.xml)

QA: No functional change, but ensure the Sign In/Sign Up/My Account/Sign Out
links continue to function properly when logging in and out.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>