BUG: The very first request in the Rails app has no per-request cache to the core server.
So this is a followup to f080ca5d6a491a1ddf4fa4aaf791546fc041db75. As it
turns out, the fix there doesn't work entirely either, because the very
first request has no per-request caching enabled, which pushes us up to
as many as 15 requests on dataset pages (vs 8 normally).

What's going on is that the RequestStore cache store doesn't get loaded
up when Rails starts up, because Rails' initializer only loads by
default the configured cache store, which is Memcache. This means that
at the start of the request, when the Rack middleware for LocalCache is
loading up MemoryStores, it doesn't load one up for RequestStore. Then,
RequestStore is finally initialized and registered with LocalCache, but
it's too late, so grabbing the active cache just returns nil.

The quick and dirty solution: initialize RequestStore once when Rails
loads up, then throw away that instance.

QA: Difficult to test, but you shouldn't have the dead-stop error on
dataset load anymore.

Reviewed-by: Paul Paradise <paul.paradise@socrata.com>