FEATURE: Adding some accesskeys in critical spots for 508
QA: pressing alt+c should skip to the content (if the page is long enough to scroll), alt+f should skip to the footer links. pressing alt+d on the alt dataset page should skip to the data, and alt+m should skip to the metadata. alt+h on any page should take you to the homepage.

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>