BUG: Separate form from create
Because the form submit requires the HTTP response object it can't be
part of a batch request. For form submit this makes sense, but since
form submit and row create were the same method it also meant that row
create couldn't be part of a batch request. Bad voodoo.

QA: Make sure form submit still works.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>