BUILD: gem bundle support
Rails 2.x added support for declaring your gem dependencies in
config/environment.rb, and added some rake tasks that allow you to
install missing gems. Unfortunately, this solution isn't particularly
ideal because by the time you get Rails running that initializer code,
you likely already depend on a gem that's not installed - so you crash
before getting a chance to bundle those gems.

Rails 3.0 is introducing a standalone gem (called 'bundler') that runs
independent of Rails to do dependency checking and maintenance of a
locked virtual environment for your dependent gems. Using this will
allow me to setup Hudson so that it bundles the gems as part of the
rails testing, rather than remembering to pre-install all the gems on
each Hudson build host/slave. With some capistrano magic, it also makes
deploying gems with native extensions a non-event, since we couldn't
vendor those before. (Admittedly, the old-style config.gem syntax would
usually work for that...)

After this checkin, the first time you run it you'll need to run:
  sudo gem install bundler
  cd web
  bundle install
In the future, if someone changes the Gemfile, you'll only need to run:
  bundle install

QA: No feature changes, but the way we load some components in the
webapp changed. Check things that we pull in as libraries, such as the
visualization code, and multipart uploads (nominations with attachments)

Reviewed-by: Chris Metcalf <chris.metcalf@socrata.com>