BUG: Sharing with a user (not email) wasn't working
The underlying problem is that we were serializing both the userId and
email regardless of which type of share we were creating, and when we
serialized an email address, the userId came out as "&nbsp;" - which we
then tried to look up, and failed.

QA: Sharing with an email address and userId should work. Kevin reproed
this earlier today by trying to send a share with a lot of users. In
theory you should be able to only repro it by just trying to share to a
contact, rather than an email address, but I didn't necessarily get a
consistent repro - I think you have to add a user share first, then add
an email share on the same dialog.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>