BUG: Crash viewing paged alt grid data
In the past few days, we started getting some Hoptoad errors along the
lines of:
  NoMethodError: undefined method '-' for "3":String

This was because we were treating params[:page] as an int without
properly casting it to be one.

QA: Typically this was only seen by search engines and spiders, but URLs
of the form:
  http://domain/?sort_by=CATEGORY&type=all&page=123
were failing to load and displaying a Hoptoad error.

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>