BUG: 990 validation message on signup accidentally read [Object object]
Somewhere along the line, our validation code broke (I suspect when we
last upgraded jQuery) such that the validation code runs a second time
onfocusout, and the arguments to the validation routine are different.
The real fix appears to be somewhere deep in the bowels of the jQuery
validator plugin I couldn't trace down; the cheap solution is to just
not use variable replacement in the validation message.

QA: https://redmine.socrata.com/issues/990

Reviewed-by: Jeff Scherpelz <jeff.scherpelz@socrata.com>