FEATURE: SDP Refresh Customizer
QA: Okay, if you go to a dataset and then go to the advanced
publisher from there, it will now work only with the new widget.

If you try to load one of the old themes, it should give you an error
message and tell you that you have to upgrade it before you can continue
to use it; it should also remove all the edit features while the error
message is visible so that you cannot edit it even by accident. You then
can either select another theme from the dropdown (and if it's a new theme
the interface will unlock), or press the button to convert the theme to
the new style format. It will copy over as much as it can about the colors
and everything, but it likely won't look very good. Oh well.

If an option used to exist but does not anymore, we don't support it
anymore. The only exception is the menu button color, which I haven't done
yet, but will eventually be customizable.

There is a known bug where if you remove every single menu item the styling
is broken. I'll fix it on Tuesday.

Reviewed-by: Aiden Scandella <aiden.scandella@socrata.com>